test_formats only contained ICLFormat tests. Migrated these to test_operators, now performed by ModelInputFormatter, and deleted file test_formats.py
